axi_logic_analyzer: Add sync to ADC path mechanism
The number of delay taps in the path can be controlled manually, from the
regmap or automatically, according to the axi_adc_decimate's rate.

Moreover, because the rate is configure by software, and the time of
initialization, is different for the ADC path and LA path. There is an
uncertainty of plus/minus one sample between the two. Because ADC and LA
paths share the same clock we can easily synchronize the two paths. We
can't use reset, because the rate generation mechanism is different
between the two. So the ADC path is used as master valid generator and we
synchronize the LA path to it.
The synchronization is done by setting the rate sync bit, which will be
cleared after the sync is achieved.